Add initial implementation of SwordController#service_document
Needs more work. Need to remove the hard coded "test" case.
Also, after updating the Collection and Depositor models
(need to verify if the Depositor really needs more attributes),
and after creating the has_and_belongs_to_many relationship,
need to update the helper SwordHelper#service_document_xml
When that is done, update #service_document to pass in the
correct arguments to the helper, and set instance vars if needed.