Add project to depositor_collection_permissions table
Need to rename this table in the future, since it's not just stating
if the depositor has access to the sword collection, it also gives,
right now, the project this is the destination for this deposit.
In our current setup, the destination is the project in hyacinth.
One possible option for rename:
depositor_collection_destinations . Sounds weird, though. Maybe repo?
Maybe ingest? ingest_info? Then the table can contain the name of the
ingest class, and a hash of options specific to that ingest class.