Add functionality to Hyacinth composer and ingest
There is now a test which calls HyacinthIngest#ingest_json.
However this test needs a live Hyacinth dev instance to respond
to the above call. So, for now, the call is commented out. Need
to stub this out, or add a conditional to check for live Hyacinth
dev instance on local machine.