Skip tests that require an active instance of hyacinth
A couple of the tests will actually make a deposit via the sword
controller. These tests will fail if there is not an active
instance of hyacinth ready to accept the contents of the deposit.
These tests are useful, so they should be kept in the code base,
but they are marked as skipped.