Add guard clause for note and subjects when generating hyacinth JSON
Previous version of the code would create an empty entry for note
in the JSON sent to hyacinth if there was no note metadata in the
parsed mets.xml file. However, it makes more sense to just not
create an entry at all.
Similar behavior was changed for subject_topic and
subject_geographic.
Unrelated changes: unused code was removed, and a new test case
was added to the mets-to-json integration test for Tow Journalism.