Remove redundant Abstract word from SpringerNature-supplied abstract
The abstract value supplied in the mets.xml file generated by
Springer Nature starts out with whitespace and the word Abstract.
Catalogers found this word redundant, so it is being removed (as well
as the preceding whitespace and the following line return) when the
abstract value is being parsed out from the mets.xml file.