Handle case where Springer Nature mets.xml has missing metadata
In the first batch of deposits, server returned 500 error
for a few of the deposit attempts. Looking at the log file, error
due to calling method on nil object. According to the error, the
mets.xml file either did not have an entry for the abstract or
the bibliographic metadata.
Now, the code will just return nil if the expected associated
xml element is not in the mets.xml file